% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_alerts.R
\name{alert_error}
\alias{alert_error}
\title{Error alert}
\usage{
alert_error(text = "")
}
\arguments{
\item{text}{the label to show in the alert}
}
\value{
a shinyalert
}
\description{
Thin wrapper around shinyalert for showing error alerts
}
