% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{findlines}
\alias{findlines}
\title{findlines}
\usage{
findlines(tempdata, searchterms)
}
\arguments{
\item{tempdata}{A list of characters to be searched.}

\item{searchterms}{Strings to search for.}
}
\value{
A list of indicies.
}
\description{
Returns indicies of items of a list that contain searchterms.
}
\examples{
findlines(c("text", "test"), "text")
}
