% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_script}
\alias{import_script}
\title{Convert a word document script into a dataframe}
\usage{
import_script(path)
}
\arguments{
\item{path}{the path to the script}
}
\value{
a single dataframe with one row per line
}
\description{
Take in the path to a word document and transform the script into a data
frame where each row is a line with a speaker.
}
