% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_01_landing.R
\name{landing_server}
\alias{landing_server}
\title{Landing page module server}
\usage{
landing_server(id, data)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{data}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Landing page module
}
