% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_01_landing.R
\name{landing_ui}
\alias{landing_ui}
\title{Landing page module
Landing page module UI}
\usage{
landing_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Landing page module
}
