% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the app}
\usage{
launch_app()
}
\value{
either run the app as a side effect or return a shiny.appobj object
}
\description{
Initiate an instance of the ScriptSearch app.
}
\details{
Note: an alias function 'l' is also provided to speed up workflow (after making changes, load them with 'Ctrl+Shift+L',
then launch the app with 'l()')
}
