% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script.R
\name{Script}
\alias{Script}
\title{Script R6 class}
\value{
returns an R6 object of class Script
}
\description{
R6 class for storing information about a script
and methods for analysing the text
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{The file path to the script}

\item{\code{name}}{The title of the script/play}

\item{\code{raw_text}}{The raw, unaltered text extracted from the file}

\item{\code{text_df}}{The raw text, formatted as a dataframe with one
row per line}

\item{\code{search_results}}{A subset of text_df, filtered using
the $search() method}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Script$new()}}
\item \href{#method-print}{\code{Script$print()}}
\item \href{#method-search}{\code{Script$search()}}
\item \href{#method-clone}{\code{Script$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new Script object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$new(path, title)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{The file path to the script}

\item{\code{title}}{(optional) the title of the script}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an R6 object with class Script
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print info about script
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-search"></a>}}
\if{latex}{\out{\hypertarget{method-search}{}}}
\subsection{Method \code{search()}}{
Search for a string and return lines that match the string
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$search(string)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{string}}{the string to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a subset of the text_df dataframe with matching lines
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Script$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
