% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_search.R
\name{search_server}
\alias{search_server}
\title{Search module server}
\usage{
search_server(id, data)
}
\arguments{
\item{id}{to be used to namespace the module}

\item{appData}{the appData reactiveValues object, defined in server.R and available
to all modules}
}
\value{
the module server, returned from a moduleServer function call
}
\description{
Defines the server logic for the Search module
}
