% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_02_search.R
\name{search_ui}
\alias{search_ui}
\title{Search module
Search module UI}
\usage{
search_ui(id)
}
\arguments{
\item{id}{to be used to namespace the module}
}
\value{
the module UI, as a tagList
}
\description{
Defines the UI for the Search module
}
