% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchdir.R
\name{searchdir}
\alias{searchdir}
\title{searchdir}
\usage{
searchdir(dir, filetypes)
}
\arguments{
\item{dir}{Path of the directory.}

\item{filetypes}{Strings to be recognised in file names.}
}
\value{
List of files.
}
\description{
Lists files in a directory.
}
\examples{
searchdir(".", "\\\\.R$")
}
