% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortenstring.R
\name{shortenstring}
\alias{shortenstring}
\title{shortenstring}
\usage{
shortenstring(string)
}
\arguments{
\item{string}{The string to be shortened.}
}
\value{
The shortened string.
}
\description{
Shortens a string if it is longer than 50 characters.
}
\examples{
shortenstring("text")
}
