% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showlines.R
\name{showlines}
\alias{showlines}
\title{showlines}
\usage{
showlines(data, n, open = TRUE)
}
\arguments{
\item{data}{The data frame output from scriptsearch().}

\item{n}{The index of the file to be searched.}

\item{open}{Logical option for opening the selected file.}
}
\value{
Table with the line number and first 50 characters of each line that contains searchterms.
}
\description{
Open and list lines containing searchterms from the nth file in the output of scriptsearch().
}
\examples{

data <- scriptsearch(".", "text")
showlines(data, 1, open = TRUE)
}
