% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptsearch.R
\name{scriptsearch}
\alias{scriptsearch}
\title{scriptsearch}
\usage{
scriptsearch(
  searchterms,
  dir = ".",
  rm = TRUE,
  filetypes = c("R", "Rmd"),
  escape = TRUE
)
}
\arguments{
\item{searchterms}{String/ list of strings to search for.}

\item{dir}{Path to the directory.}

\item{rm}{Logical option for removing files with no hits from the output.}

\item{filetypes}{String/list of strings to be recognised in file endings.}

\item{escape}{Logical option for adding escape characters to special characters.}
}
\value{
Data frame containing the file paths, number of hits and search terms found in a directory.
}
\description{
Returns the number of times a searchterm appears in each file in a path (defaults to working directory).
}
\examples{
scriptsearch("text")

scriptsearch("text", filetypes = "txt")

data <- scriptsearch("text", rm = FALSE)
}
