% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptsearch.R
\name{scriptsearch}
\alias{scriptsearch}
\title{scriptsearch}
\usage{
scriptsearch(
  dir,
  searchterms,
  rm = TRUE,
  filetypes = c("\\\\.R$", "\\\\.Rmd$")
)
}
\arguments{
\item{dir}{Path of the directory.}

\item{searchterms}{Strings to search for.}

\item{rm}{TRUE/FALSE for removing files with no hits from the output.}

\item{filetypes}{Strings to be recognised in file names.}
}
\value{
Data frame containing: file paths, TRUE/FALSE if text detected and search terms.
}
\description{
Create a table listing files in a directory that contain specified strings.
}
\examples{
scriptsearch(".", "x")
}
