% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showlines.R
\name{showlines}
\alias{showlines}
\title{showlines}
\usage{
showlines(data, n, open = FALSE)
}
\arguments{
\item{data}{The data frame output from scriptsearch().}

\item{n}{The row number of the file to be searched.}

\item{open}{Logical option for opening the selected file.}
}
\value{
Table with the line number and first 50 characters of each line that contains searchterms.
}
\description{
List the lines containing searchterms from the file in the nth row of a data frame (output of scriptsearch).
}
\examples{
showlines(scriptsearch(".", "search"), 1, open = TRUE)
}
